backend/requirements.txt
Django>=3.0,<4.0
djangorestframework
mysqlclient
django-cors-headers